+++ source/videostreamlibuvc.h
@@ -25,6 +25,7 @@ public:
     ~VideoStreamLibUVC() override;
 
     void setBufferParameters(cv::Mat *frameBuf, qint64 *tsBuf, float *bnoBuf,
+                             qint64 *daqFrameNumBuf,
                              int bufferSize, QSemaphore *freeFramesS, QSemaphore *usedFramesS,
                              QAtomicInt *acqFrameNum, QAtomicInt *daqFrameNumber) override;
     int connect2Camera(int cameraID) override;
@@ -72,6 +73,7 @@ private:
 
     cv::Mat *frameBuffer;
     qint64 *timeStampBuffer;
+    qint64 *daqFrameNumBuffer;
     float *bnoBuffer;
     QSemaphore *freeFrames;
     QSemaphore *usedFrames;